+++ handlers/schedule_func.py
@@ -1,95 +1,21 @@
-import os
 from datetime import datetime, timedelta
 
 from aiogram import types, F
 from aiogram.filters import Command
 from aiogram.fsm.context import FSMContext
-from aiogram.fsm.state import State, StatesGroup
-from aiogram.types import FSInputFile
 from aiogram.utils.keyboard import ReplyKeyboardBuilder
-from PIL import Image, ImageDraw, ImageFont
-
-from database.create_tables import session
-from database.models import BookTime
-
-from loader import dp, bot, get_user
+from loader import bot, get_user
 from loader import form_router
 
-class ScheduleForm(StatesGroup):
-    askForDate = State()
-    showSchedule = State()
+from states.states import ScheduleForm
 
+from utils.image import generate_schedule_image, send_image
 
-def more_schedule_kb_for_admin():
-    builder = ReplyKeyboardBuilder()
-    builder.add(types.KeyboardButton(text="Посмотреть еще"))
-    builder.add(types.KeyboardButton(text="📌Забронировать"))
-    builder.adjust(1)
-    return builder.as_markup(one_time_kyeboard=True, resize_keyboard=True)
+from keyboards.reply.usermodes import (
+    more_schedule_kb_for_user,
+    more_schedule_kb_for_admin,
+)
 
-
-def more_schedule_kb_for_user():
-    builder = ReplyKeyboardBuilder()
-    builder.add(types.KeyboardButton(text="Посмотреть еще"))
-    builder.adjust(1)
-    return builder.as_markup(one_time_kyeboard=True, resize_keyboard=True)
-
-
-def fetch_event_for_date(date: str):
-    records = session.query(BookTime).filter_by(date=date).all()
-    if records:
-        data = []
-        for record in records:
-            data.append((record.startTime, record.endTime, record.reason, record.renter))
-        return data
-    else:
-        return ""
-
-
-async def generate_schedule_image(date, state: FSMContext):
-    text_color = (0, 0, 0)
-    font_size = 32
-    days_translation = {
-        "Monday": "Понедельник",
-        "Tuesday": "Вторник",
-        "Wednesday": "Среда",
-        "Thursday": "Четверг",
-        "Friday": "Пятница",
-        "Saturday": "Суббота",
-        "Sunday": "Воскресенье"
-    }
-    image = Image.open("handlers/images/бронь_120.jpeg", mode="r").convert("RGB")
-    draw = ImageDraw.Draw(image)
-    font = ImageFont.truetype(
-        "handlers/fonts/DejaVuSans-Bold.ttf", font_size, encoding="unic"
-    )
-
-    day = date.strftime("%d.%m")
-    await state.update_data(day=day)
-    data = fetch_event_for_date(day)
-    day_of_week = days_translation[date.strftime("%A")]
-
-    draw.text((50, 50), f"{day_of_week} {day}", fill=text_color, font=font)
-    y_offset = 110
-    font_size = 25
-    font = ImageFont.truetype(
-        "handlers/fonts/DejaVuSans.ttf", font_size, encoding="unic"
-    )
-
-    if isinstance(data, list):
-        for startTime, endTime, reason, renter in data:
-            draw.text(
-                (50, y_offset),
-                f"{startTime} - {endTime}\nТема: {reason}",
-                fill=text_color,
-                font=font,
-            )
-            y_offset += 77
-    else:
-        draw.text((10, y_offset), data, fill=text_color, font=font)
-    image_path = "handlers/images/schedule.jpg"
-    image.save(image_path)
-    return image_path
 
 
 @form_router.message(Command("schedule"))
@@ -112,6 +38,7 @@ async def schedule_command(message: types.Message, state: FSMContext):
     await state.update_data(
         last_user_message=message.message_id, last_bot_message=bot_message.message_id
     )
+
 
 @form_router.message(ScheduleForm.askForDate)
 async def process_date_selection(message: types.Message, state: FSMContext):
@@ -137,26 +64,6 @@ async def process_date_selection(message: types.Message, state: FSMContext):
     user = await get_user(message)
 
     if user.status == "left":
-        if os.path.exists(photo_path):
-            photo = FSInputFile(photo_path)
-            await message.answer_photo(photo, caption=f"Расписание на {selected_date}", reply_markup=more_schedule_kb_for_user())
-            records = fetch_event_for_date(day)
-            for startTime, endTime, reason, renter in records:
-                await message.answer(f"Время: {startTime}-{endTime}\nТема: {reason}\nАрендатор: @{renter}")
-
-        else:
-            await message.answer("Изображение с расписанием не найдено.")
-
-        await state.clear()
+        send_image(photo_path, selected_date, more_schedule_kb_for_user, day, message, state)
     else:
-        if os.path.exists(photo_path):
-            photo = FSInputFile(photo_path)
-            await message.answer_photo(photo, caption=f"Расписание на {selected_date}",
-                                       reply_markup=more_schedule_kb_for_admin())
-            records = fetch_event_for_date(day)
-            for startTime, endTime, reason, renter in records:
-                await message.answer(f"Время: {startTime}-{endTime}\nТема: {reason}\nАрендатор: @{renter}")
-        else:
-            await message.answer("Изображение с расписанием не найдено.")
-
-        await state.clear()
+        send_image(photo_path, selected_date, more_schedule_kb_for_admin, day, message, state)
